Updated to version A16
Added Console.pause to pause the program, uses Thread.sleep and a loop as a fail-safe.
Added prompt input verification that guarantees that the input entered will be passable to a type.
Added console print data object to store all the data need for a print to console in one object/parameter. Used in the prompt input verification.